Issue #133: Populate DataCollection Bug fix on NewPageIterator: Incorrect MailDiscriminatorValue. DataCollection implemented. Brute force. However, this does not use the populate method. The populate method to be added in the next step.
Test Suite: TestMailCount is an end-to-end test. Tests to see if the
amount of values in a mailbox are the amount of messages that are
retrieved. This is false at present. There are several errors from
ParseNodePackage: `unsupported AdditionalData type:
map[string]*jsonserialization.JsonParseNode`. Will have to determine how
errors are to be sent to the DataCollection. Issue #137. Upon
inspection, there are 59 messages on the test account. 7 of them receive
a parsing error and 2 are unaccounted for. Issue #138