add no-op action for required jobs
Adds a no-op job for Test-Suite and Linting.
According to github docs, this will mark those
jobs as successful unless otherwise failed.  So on
a run that doesn't trigger the standard workflows
the job will always succeed.  On runs that trigger
workflows, it will succeed until overwritten with
a failure.