Turn issues into bullets so they show titles (#754)
## Description

Minor visual change to how issues for the PR are displayed. Now defaults to the bulleted list which causes GitHub to render the title of each issue instead of just the issue number. Issues are still links

## Type of change

Please check the type of change your PR introduces:
- [ ] :sunflower: Feature
- [ ] :bug: Bugfix
- [ ] :world_map: Documentation
- [ ] :robot: Test
- [x] :hamster: Trivial/Minor

## Issue(s)
<!-- Can reference multiple issues. Use one of the following "magic words" - "closes, fixes" to auto-close the Github issue. -->

As an illustration, the below does not use bullets:
#753 

This does use bullets though
* closes #753 

## Test Plan

<!-- How will this be tested prior to merging.-->

- [x] :muscle: Manual
- [ ] :zap: Unit test
- [ ] :green_heart: E2E