slow load test frequency to every 4 hours
load testing for the dev m356 tenant takes longer
than 3 hours to complete, causing all load tests to
get canceled when github shifts priority to the
next job that gets kicked off.  Backing off the freq
to every 4 hours, and will revisit that as needed.
If 4 hour time doesn't satisfy the boundary, we
should shift the dataset instead of freqnency.