add deets to gc restore ops (#1029)
## Description

In order for a restore operation to accurately print
out the list of items that were restored, it needs to
track item details within the GC restoration like is
done within kopia in backup.   These details will
not be stored in a modelStore.  But they will get
printed out on the CLI for end users.

Part 1 of multiple.

## Type of change

- [x] :sunflower: Feature

## Issue(s)

* #977

## Test Plan

- [x] :muscle: Manual
- [x] :green_heart: E2E