Don't store stable ID in serialized JSON
Stable ID is used as a lookup key which means the model store stores it
in the tags. Due to that it doesn't need to be serialied into the JSON
of the struct as well. Code that reads the model will repopulate the
field properly prior to returning the read data to the caller.