Add helper function to get drive ID
Create a OneDrive helper function to get a drive ID for the user's
default drive. It needs to be exported but can't be placed in a _test.go
or testdata directory because those either don't export or lint
properly. Instead, add a build constraint 'testing' which will keep it
from being built as part of the prod binary. Unfortunatly this also
means that one must use go test -tags testing to properly compile tests
using this function.