adds collection-level progress display
Adds a progress spinner tracking the collection-level
work on an incremental basis.  Useful for tracking
collection progress when individual files may be too
quickly resolved to visibly show progress bars.

In the future, we could combine the two so that the
current item in progress always displays below the
collection tracker.