adds collection-level progress display (#1187)
## Description

Adds a progress spinner tracking the collection-level work on an incremental basis.  Useful for tracking collection progress when individual files may be too quickly resolved to visibly show progress bars.

In the future, we could combine the two so that the current item in progress always displays below the collection tracker.

## Type of change

- [x] :sunflower: Feature

## Issue(s)

* #1112

## Test Plan

- [x] :muscle: Manual
- [x] :zap: Unit test