Add a make clean step for the website
`make clean` will now be executed before a Docker image is built.
While it technically has no impact on the Docker image, it will
ensure that there are no stale Node modules lying around when a
build is executed after the creation of a new image.

While running a Docker build without a cache is an additional option,
it does not prevent stale local node_modules directories and suffers
from the overhead of having to pull down a large Ubuntu base image
every time.