Split multi-tuple backups into corso backups
Split backups that span multiple (resource owner, service, category)
tuples into discrete corso backups, one per tuple. This shards the
backup details and allows for better parallelism in the future if we
want it. Could also be useful for incremental backups.

Some things that don't fit well with the current patterns in the code:
* same selector saved in each backup making it difficult to determine
  what the backup contains from the summary information
* number of items read by GC is aggregate for all backups so is no
  longer persisted
* many backup IDs now need to be reported to caller
* status handling for the operation vs. individual backups