Create MetadataCollection
MetadataCollection is a collection that buffers data until it is read.
It is not meant to store large amount of data since all data is resident
in-memory.

Items in this collection do not implement StreamInfo because it is not
intended for use by things that should be surfaced to the user through
backup details.