Relax some interface checks in KopiaWrapper
The delta token file does not implement StreamInfo because it is not
something exposed to the user. Relax the checks in KopiaWrapper to
account for this. Files that do not have info added to the progress
tracker in KopiaWrapper will be ignored when kopia informs Corso it's
done with the file via the callback.