Add function to resurface items from prev snapshot
Function merges items from a previous snapshot with items from the
collection for the current snapshot. Items that have already been seen
(according to item UUID) are skipped in the base snapshot to avoid
ending up with duplicate items. Subdirectories of the base snapshot's
directory are not traversed in this function.