Test helper functions for checking hierarchies
Create some helper functions capable of walking a tree of nodes and
checking that the given input fs.Entry hierarchy contains entries with
the proper names. Can also check file data if that is specified.
However, the input files are expected to have Corso serialization
version identifiers at the start.