Update helper functions to propagate reasons
Add and propagate reason for selecting a snapshot while looking for the
most recent snapshots.

This code can be simplified and optimized once Corso uses a more recent
version of Kopia that allows adding tags to checkpoints. The logic to
find complete/incomplete snapshots will become much simpler because
inspection of the tags will allow determining if a snapshot is complete.