Generate path based on Reason snapshot selected
Use information such as the root directory name and reason a snapshot
was selected to generate the path the subtree of data we're intested in
is at.

This code could be moved to BackupOp and may be better suited there
because it does pull in quite a bit of extra context/assumptions to the
kopia package.