Rewrite helper functions some to be more generic
Make helper functions for generating expected and base snapshot trees to
be more generic. Instead of taking a fixed set of parameters for stuff
like tenant, user, service, etc, take a list. The items in the list will
be consumed in order and added to the tree. Each item will be the only
child of the previous item in the tree.

Allow giving a more complex tree at the very bottom level via the
previous `children` parameter. That represents the remainder of the tree
after the other items have been inflated.