Fetch isRead attribute in email delta queries
Updating the isRead flag in M365 does not cause the mod time of the item
to be updated. This means delta queries will not return items marked as
[un]read when isRead is not one of the attributes requested in the
$select clause.

Explicitly fetch isRead so that we get those items. Further changes will
be required to ensure the mod time of the item is something that will
not cause kopia-assisted dedupe to skip the item when only the isRead
attribute has been changed.